app/requirements.txt
databricks-sdk>=0.20.0
databricks-sql-connector>=3.0.0
streamlit>=1.35.0
pandas>=2.0.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
